+++ examples/basic.rs
@@ -34,7 +34,8 @@ fn main() {
     future::block_on(async {
         let work = async {
             let sum = ticks
-                .progress_with_messages(theme, |index, _| index as f64 / 100.0, messages)
+                .progress(theme, |index, _| index as f64 / 100.0)
+                .with_messages(messages)
                 .fold(0, |acc, x| acc + x)
                 .await;